requirements.txt
-i https://pypi.org/simple
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.1
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0